requirements.txt
pandas==1.5.2
Pillow==9.4.0
plotly==5.11.0
psycopg2==2.9.5
python_bcrypt==0.3.2
PyYAML==6.0
SQLAlchemy==1.4.46
streamlit==1.16.0
streamlit_authenticator==0.2.1
streamlit_autorefresh==0.0.1
